FastAPI/pyproject.toml
[tool.poetry]
name = "calorie-tracker-backend"
version = "0.1.0"
description = "FastAPI backend for calorie tracking"
authors = ["Maciej <maciej.trznadel24@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "0.116.1"
alembic = "1.16.5"
email-validator = "2.3.0"
psycopg2-binary = "2.9.10"
pydantic = "2.11.7"
python-dotenv = "1.1.1"
sqlalchemy = "2.0.43"
uvicorn = "0.35.0"
python-jose = "3.5.0"
passlib = "1.7.4"
bcrypt = "5.0.0"
python-multipart = "0.0.20"
zxcvbn = "^4.5.0"
redis = "^7.0.0"
pydantic-settings = "^2.12.0"
asyncpg = "^0.30.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
pytest-cov = "^7.0.0"
httpx = "^0.28.1"
ruff = "^0.14.0"
black = "^25.9.0"
isort = "^7.0.0"
pre-commit = "^4.3.0"
fakeredis = {extras = ["aioredis"], version = "^2.32.1"}


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "-ra",
]

markers = [
    "integration: integration tests",
    "unit: unit tests",
    "e2e: end to end tests"
]


[tool.coverage.run]
source = ["app"]
omit = [
    "*/__init__.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = true


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
